pyproject.toml
[project]
name = "chatbot-telegram"
version = "0.1.0"
description = "Este projeto foi desenvolvido como parte de um teste técnico, para criar uma experiência conversacional útil e engajadora para a torcida, por Matheus Calegari."
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "beautifulsoup4==4.13.4",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.1",
    "colorama==0.4.6",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.25.0rc0",
    "google-api-python-client==2.169.0",
    "google-auth==2.39.0",
    "google-auth-httplib2==0.2.0",
    "google-generativeai==0.8.5",
    "googleapis-common-protos==1.70.0",
    "grpcio==1.71.0",
    "grpcio-status==1.71.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "idna==3.10",
    "proto-plus==1.26.1",
    "protobuf==5.29.4",
    "pyasn1==0.6.1",
    "pyasn1_modules==0.4.2",
    "pydantic==2.11.4",
    "pydantic_core==2.33.2",
    "pyparsing==3.2.3",
    "python-dotenv==1.1.0",
    "python-telegram-bot==22.0",
    "requests==2.32.3",
    "rsa==4.9.1",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "tqdm==4.67.1",
    "typing-inspection==0.4.0",
    "typing_extensions==4.13.2",
    "uritemplate==4.1.1",
    "urllib3==2.4.0"
]
